backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
Pillow==9.3.0
djangorestframework-simplejwt==4.7.2
django-filter==21.1
